backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.2,<6.0
psycopg2-binary>=2.9,<3.0
openai>=1.0,<2.0
requests>=2.31,<3.0
celery>=5.3,<6.0
redis>=4.5,<5.0
